gcp/requirements.txt
tensorflow
google-cloud-storage
pillow

